requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
imagehash>=4.3.1
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.0
watchdog>=3.0.0
gspread>=5.10.0
oauth2client>=4.1.3
telethon>=1.30.0
